+++ src/f5cmisc.cuh
@@ -48,13 +48,15 @@ __global__ void
 align_kernel_core_2d_shm(int32_t* read_len, int32_t* read_ptr,
     event_t* event_table, int32_t* n_events1,
     int32_t* event_ptr, 
-    scalings_t* scalings, int32_t n_bam_rec,model_t* model_kmer_caches,float *band,uint8_t *traces, EventKmerPair* band_lower_lefts) ;
+    scalings_t* scalings, int32_t n_bam_rec, model_t* model_kmer_caches,
+    half *band,uint8_t *traces, EventKmerPair* band_lower_lefts) ;
 
 __global__ void align_kernel_pre_2d(char* read,
     int32_t* read_len, int32_t* read_ptr,
     int32_t* n_events,
     int32_t* event_ptr, model_t* models,
-    int32_t n_bam_rec,model_t* model_kmer_caches,float *bands1,uint8_t *trace1, EventKmerPair* band_lower_left1) ;
+    int32_t n_bam_rec,model_t* model_kmer_caches, half *bands1,
+    uint8_t *trace1, EventKmerPair* band_lower_left1) ;
 
 
 __global__ void align_kernel_post(AlignedPair* event_align_pairs,
@@ -62,7 +64,8 @@ __global__ void align_kernel_post(AlignedPair* event_align_pairs,
     int32_t* read_len, int32_t* read_ptr,
     event_t* event_table, int32_t* n_events,
     int32_t* event_ptr, 
-    scalings_t* scalings, int32_t n_bam_rec,model_t* model_kmer_caches,float *bands1,uint8_t *trace1, EventKmerPair* band_lower_left1);
+    scalings_t* scalings, int32_t n_bam_rec,model_t* model_kmer_caches,
+    half *bands1,uint8_t *trace1, EventKmerPair* band_lower_left1);
